bovista::visual

Trait Visual

pub trait Visual:
    Send
    + Sync
    + Any {
    // Required methods
    fn prepare(
        &mut self,
        device: &Device,
        queue: &Queue,
        camera_info: &CameraInfo,
    );
    fn render(&self, render_pass: &mut RenderPass<'_>);
    fn set_transform(&mut self, transform: Transform);
    fn transform(&self) -> &Transform;
    fn is_visible(&self) -> bool;
    fn set_visible(&mut self, visible: bool);

    // Provided method
    fn name(&self) -> &str { ... }
}

Base trait for all visual objects that can be rendered in a scene

Required Methods

fn prepare(&mut self, device: &Device, queue: &Queue, camera_info: &CameraInfo)

Prepare the visual for rendering (called once per frame before render) This is where uniform buffers should be updated, etc. Camera info is provided for visuals that need LOD selection or culling

fn render(&self, render_pass: &mut RenderPass<'_>)

Render the visual to the given render pass Note: This method cannot take ’self by reference with a lifetime because it needs to work with Mutex guards in the scene graph

fn set_transform(&mut self, transform: Transform)

Update the visual’s transform matrix

fn transform(&self) -> &Transform

Get the current transform

fn is_visible(&self) -> bool

Check if the visual is visible

fn set_visible(&mut self, visible: bool)

Set visibility

Provided Methods

fn name(&self) -> &str

Get a human-readable name for this visual (for debugging)